Refactored secret fetcher actors
