+++ ydb/core/kqp/executer_actor/ya.make
@@ -27,7 +27,7 @@ PEERDIR(
     ydb/core/kqp/common
     ydb/core/kqp/compute_actor
     ydb/core/kqp/executer_actor/shards_resolver
-    ydb/core/kqp/federated_query
+    ydb/core/kqp/federated_query/actors
     ydb/core/kqp/gateway/local_rpc
     ydb/core/kqp/query_compiler
     ydb/core/kqp/rm_service
