+++ ydb/core/kqp/federated_query/actors/ut_service/common/ya.make
@@ -8,7 +8,7 @@ PEERDIR(
     ydb/public/sdk/cpp/src/client/table
     ydb/library/aclib
     ydb/core/kqp/common/events
-    ydb/core/kqp/federated_query
+    ydb/core/kqp/federated_query/actors
     ydb/core/kqp/ut/common
 )
