+++ ydb/core/kqp/federated_query/actors/ut_service/fast/kqp_federated_query_actors_ut.cpp
@@ -1,8 +1,7 @@
-#include "kqp_federated_query_actors.h"
-
 #include <ydb/core/kqp/common/events/script_executions.h>
 #include <ydb/core/kqp/common/simple/services.h>
-#include <ydb/core/kqp/federated_query/ut_service/common/helpers.h>
+#include <ydb/core/kqp/federated_query/actors/kqp_federated_query_actors.h>
+#include <ydb/core/kqp/federated_query/actors/ut_service/common/helpers.h>
 #include <ydb/core/kqp/ut/common/kqp_ut_common.h>
 
 #include <library/cpp/testing/unittest/registar.h>
