+++ ydb/core/kqp/federated_query/actors/ut_service/fast/ya.make
@@ -1,10 +1,10 @@
-UNITTEST_FOR(ydb/core/kqp/federated_query)
+UNITTEST_FOR(ydb/core/kqp/federated_query/actors)
 
 SIZE(MEDIUM)
 
 PEERDIR(
-    ydb/core/kqp/federated_query
-    ydb/core/kqp/federated_query/ut_service/common
+    ydb/core/kqp/federated_query/actors
+    ydb/core/kqp/federated_query/actors/ut_service/common
     ydb/core/kqp/ut/common
     yql/essentials/sql/pg_dummy
 )
