+++ ydb/core/kqp/federated_query/actors/ut_service/ya.make
