+++ ydb/core/kqp/federated_query/ya.make
@@ -1,16 +1,13 @@
 LIBRARY()
 
 SRCS(
-    kqp_federated_query_actors.cpp
     kqp_federated_query_helpers.cpp
 )
 
 PEERDIR(
     ydb/core/base
-    ydb/core/fq/libs/grpc
     ydb/core/fq/libs/db_id_async_resolver_impl
     ydb/core/protos
-    ydb/library/db_pool/protos
     ydb/library/logger
     ydb/library/yql/providers/common/http_gateway
     ydb/library/yql/providers/generic/connector/libcpp
@@ -33,7 +30,10 @@ YQL_LAST_ABI_VERSION()
 
 END()
 
+RECURSE(
+    actors
+)
+
 RECURSE_FOR_TESTS(
     ut
-    ut_service
 )
